Recipe/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.5
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
idna==2.10
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
